// server/prisma/generated/enums.ts
export const PaymentMethod = {
  INSTORE_CASH: 'INSTORE_CASH',
  INSTORE_CARD: 'INSTORE_CARD',
  CASH_APP: 'CASH_APP'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN',
  VIP: 'VIP',
  MODERATOR: 'MODERATOR',
  SYSTEM: 'SYSTEM',
  OWNER: 'OWNER',
  MEMBER: 'MEMBER',
  OPERATOR: 'OPERATOR',
  SUPPORT_AGENT: 'SUPPORT_AGENT'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const KeyMode = {
  read: 'read',
  write: 'write',
  upload: 'upload',
  manage_users: 'manage_users',
  manage_settings: 'manage_settings',
  launch_game: 'launch_game'
} as const

export type KeyMode = (typeof KeyMode)[keyof typeof KeyMode]